Rename test_orbit.py to orbit.py to avoid pytest confusions
